adding lab looks good

+++ internal/handler/server.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"actlabs-hub/internal/auth"
 	"actlabs-hub/internal/entity"
 	"net/http"
 
@@ -24,9 +25,15 @@ func NewServerHandler(r *gin.RouterGroup, serverService entity.ServerService) {
 }
 
 func (h *serverHandler) GetServer(c *gin.Context) {
+
 	server := entity.Server{}
 	if err := c.ShouldBindJSON(&server); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	if !auth.VerifyUserObjectId(server.UserPrincipalId, c.GetHeader("Authorization")) {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid request"})
 		return
 	}
 
@@ -46,6 +53,11 @@ func (h *serverHandler) DeployServer(c *gin.Context) {
 		return
 	}
 
+	if !auth.VerifyUserObjectId(server.UserPrincipalId, c.GetHeader("Authorization")) {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid request"})
+		return
+	}
+
 	server, err := h.serverService.DeployServer(server)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -62,6 +74,11 @@ func (h *serverHandler) DestroyServer(c *gin.Context) {
 		return
 	}
 
+	if !auth.VerifyUserObjectId(server.UserPrincipalId, c.GetHeader("Authorization")) {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid request"})
+		return
+	}
+
 	err := h.serverService.DestroyServer(server)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
@@ -73,6 +90,11 @@ func (h *serverHandler) DestroyServer(c *gin.Context) {
 
 func (h *serverHandler) UpdateActivityStatus(c *gin.Context) {
 	userPrincipalName := c.Param("userPrincipalName")
+
+	if !auth.VerifyUserPrincipalName(userPrincipalName, c.GetHeader("Authorization")) {
+		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid request"})
+		return
+	}
 
 	if err := h.serverService.UpdateActivityStatus(userPrincipalName); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
